class-33/demo/live-authenticated-things/pyproject.toml
[tool.poetry]
name = "live-authenticated-things"
version = "0.1.0"
description = ""
authors = ["Dario Thornhill <dariothornhill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
